+++ components/mdc/Drawer/Drawer.svelte
@@ -15,7 +15,6 @@ export let modal = false
 export let dismissible = false
 export let toggle = false
 export let isFullHeightMenu = true
-export let classForContent = '' //can be used to change background color
 
 let mdcDrawer = {}
 let element = {}
@@ -125,7 +124,7 @@ main {
     </TopAppBar>
   {/if}
 
-  <main class="h-100" class:classForContent>
+  <main class="h-100" id="main-drawer-content">
     <div class:mdc-top-app-bar--dense-fixed-adjust={hasTopAppBar} class="h-100">
       <slot />
     </div>