+++ components/mdc/Drawer/Drawer.svelte
@@ -2,9 +2,10 @@
 <script>
 import { isAboveTablet as isDesktop } from '../breakpoints'
 import { MDCDrawer } from '@material/drawer'
-import { Button, TopAppBar } from '../index'
+import Button from '../Button'
 import { beforeUrlChange } from '@roxi/routify'
 import { onMount } from 'svelte'
+import TopAppBar from '../TopAppBar'
 
 export let title = ''
 export let subtitle = ''
@@ -14,6 +15,7 @@ export let modal = false
 export let dismissible = false
 export let toggle = false
 export let isFullHeightMenu = true
+export let hideForMobile = true
 
 let mdcDrawer = {}
 let element = {}
@@ -30,7 +32,7 @@ const isMenuItemActive = (currentUrl, menuItemUrl) => currentUrl === menuItemUrl
 
 $: currentUrl = window.location.pathname
 $: toggle, toggleDrawer()
-$: !dismissible && (modal = true) //prevents error if neither is selected
+$: !dismissible && showModalDrawer() //prevents error if neither is selected
 
 $beforeUrlChange(({ url }) => {
   currentUrl = url
@@ -39,7 +41,7 @@ $beforeUrlChange(({ url }) => {
 })
 
 const showAppropriateDrawer = () => {
-  isDesktop() && !dismissible ? showStandardDrawer() : showModalDrawer()
+  isDesktop() && !dismissible ? showStandardDrawer() : hideForMobile && showModalDrawer()
 }
 const showModalDrawer = () => modal = true
 const showStandardDrawer = () => modal = false
@@ -91,7 +93,7 @@ main {
           <span class="grow-1" />
         {:else if !hide}
           {#if url && button}
-           <Button class="m-1" raised prependIcon={typeof button == 'string' ? button : null} {url} >{label}</Button>
+           <Button class="m-1" raised prependIcon={icon} {url} >{label}</Button>
           {:else if url}
             <a class="mdc-list-item" class:mdc-list-item--activated={isMenuItemActive(currentUrl, url)} href={url}
               aria-current={isMenuItemActive(currentUrl, url) ? "page" : null} tabindex={i === 0 ? 0 : undefined}>
@@ -119,11 +121,12 @@ main {
   {#if hasTopAppBar}
     <TopAppBar dense fixed bgColorIsVariant on:nav={toggleDrawer} navIconBreakpointClass={!dismissible && "hide-above-tablet"} >
       <slot name="TopAppBar"/>
+      <slot slot="actions" name="actions"/>
     </TopAppBar>
   {/if}
 
-  <main class="h-100">
-    <div class="mdc-top-app-bar--dense-fixed-adjust h-100">
+  <main class="h-100" id="main-drawer-content">
+    <div class:mdc-top-app-bar--dense-fixed-adjust={hasTopAppBar} class="h-100">
       <slot />
     </div>
   </main>