+++ CHANGELOG.md
@@ -5,8 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [2.5.0](https://github.com/silinternational/ui-components/releases/tag/v2.5.0) - 2021-08-03
 
+## [2.6.0](https://github.com/silinternational/ui-components/releases/tag/v2.6.0) - 2021-08-04
+### Added
+- hideForMobile prop and optional border-white class in Drawer
+- another slot to TopAppBar
+- main-drawer-content id and actions slot to Drawer
+
+### Fixed
+- TopAppBar adjust class removed when not needed by Drawer
+
+## [2.5.0](https://github.com/silinternational/ui-components/releases/tag/v2.5.0) - 2021-08-03
 ### Added
 - slot to TopAppBar in Drawer